Improve cdfScript capabilities
This makes it so that duplicates are allowed, at least in type lists.
This can be useful for things that want to check a list of cards for "entirely different types", by just running DIFFERENT() on their types.
Note that the actual implementation of a card still mandates no duplicates in its type list.